Decomposed from one file supplier template to its organisms and molecules whilst maintaining same functionality and output
